code cleanup: move SetLengthMods to setlengths.c
