+++ src/libespeak-ng/setlengths.h
@@ -29,6 +29,7 @@ extern "C"
 #endif
 
 void CalcLengths(Translator *tr);
+void SetLengthMods(Translator *tr, int value);
 
 espeak_ng_STATUS SetParameter(int parameter, 
 	int value,
