+++ src/libespeak-ng/translate.h
@@ -659,7 +659,6 @@ extern int dictionary_skipwords;
 
 extern int (*uri_callback)(int, const char *, const char *);
 extern int (*phoneme_callback)(const char *);
-extern void SetLengthMods(Translator *tr, int value);
 
 #define LEADING_2_BITS 0xC0 // 0b11000000
 #define UTF8_TAIL_BITS 0x80 // 0b10000000
